Separate test for count_ones
