+++ bit/test_bit.py
@@ -17,6 +17,34 @@ class TestSuite(unittest.TestCase):
         self.assertEqual(0, add_bitwise_operator(0, 0))
         self.assertEqual(10, add_bitwise_operator(10, 0))
         self.assertEqual(10, add_bitwise_operator(0, 10))
+        
+    def test_count_ones_recur(self):
+
+        # 8 -> 1000
+        self.assertEqual(1, count_ones_recur(8))
+
+        # 109 -> 1101101
+        self.assertEqual(5, count_ones_recur(109))
+
+        # 63 -> 111111
+        self.assertEqual(6, count_ones_recur(63))
+
+        # 0 -> 0
+        self.assertEqual(0, count_ones_recur(0))
+
+    def test_count_ones_iter(self):
+
+        # 8 -> 1000
+        self.assertEqual(1, count_ones_iter(8))
+
+        # 109 -> 1101101
+        self.assertEqual(5, count_ones_iter(109))
+
+        # 63 -> 111111
+        self.assertEqual(6, count_ones_iter(63))
+
+        # 0 -> 0
+        self.assertEqual(0, count_ones_iter(0))
 
 if __name__ == '__main__':
     unittest.main()
